Add postme tied to user

+++ src/handlers/postme/postme.scene.ts
@@ -28,44 +28,9 @@ export const enterPasswordScene = new Scenes.BaseScene<BotContext>(
 );
 
 enterPasswordScene.enter(async (ctx) => {
-  const editedMessage = await ctx.editMessageText(
-    'Введите пароль для доступа к Вашему чату',
-    {
-      reply_markup: {
-        inline_keyboard: [
-          [{ text: 'Выйти', callback_data: CommonActions.ExitCallback }],
-        ],
-      },
-    },
-  );
-  if (typeof editedMessage === 'object') {
-    ctx.scene.state = {
-      ...ctx.scene.state,
-      dialogMessageId: editedMessage.message_id,
-    };
-  }
-});
-
-enterPasswordScene.on(message('text'), async (ctx) => {
-  const { id: chatId } = ctx.chat;
-  const { dialogMessageId } = ctx.scene.state as SceneState;
-
-  if (!dialogMessageId) {
-    await ctx.editMessageText(WRONG);
-    await ctx.scene.leave();
-    return;
-  }
-
-  const { text } = ctx.message;
-
-  await ctx.deleteMessage();
-
-  if (text.length < 6) {
-    await ctx.telegram.editMessageText(
-      chatId,
-      dialogMessageId,
-      undefined,
-      'Пароль должен быть больше или равен 6 символам',
+  try {
+    const editedMessage = await ctx.editMessageText(
+      'Введите пароль для доступа к Вашему чату',
       {
         reply_markup: {
           inline_keyboard: [
@@ -74,14 +39,57 @@ enterPasswordScene.on(message('text'), async (ctx) => {
         },
       },
     );
-    return undefined;
+    if (typeof editedMessage === 'object') {
+      ctx.scene.state = {
+        ...ctx.scene.state,
+        dialogMessageId: editedMessage.message_id,
+      };
+    }
+  } catch (error) {
+    console.log(error);
   }
+});
 
-  ctx.session.postme = {
-    ...ctx.scene.state,
-    password: text,
-  };
-  ctx.scene.enter(PostmeScene.ConfirmPassword);
+enterPasswordScene.on(message('text'), async (ctx) => {
+  try {
+    const { id: chatId } = ctx.chat;
+    const { dialogMessageId } = ctx.scene.state as SceneState;
+
+    if (!dialogMessageId) {
+      await ctx.editMessageText(WRONG);
+      await ctx.scene.leave();
+      return;
+    }
+
+    const { text } = ctx.message;
+
+    await ctx.deleteMessage();
+
+    if (text.length < 6) {
+      await ctx.telegram.editMessageText(
+        chatId,
+        dialogMessageId,
+        undefined,
+        'Пароль должен быть больше или равен 6 символам',
+        {
+          reply_markup: {
+            inline_keyboard: [
+              [{ text: 'Выйти', callback_data: CommonActions.ExitCallback }],
+            ],
+          },
+        },
+      );
+      return undefined;
+    }
+
+    ctx.session.postme = {
+      ...ctx.scene.state,
+      password: text,
+    };
+    ctx.scene.enter(PostmeScene.ConfirmPassword);
+  } catch (error) {
+    console.log(error);
+  }
 });
 
 export const confirmPasswordScene = new Scenes.BaseScene<BotContext>(
@@ -89,37 +97,15 @@ export const confirmPasswordScene = new Scenes.BaseScene<BotContext>(
 );
 
 confirmPasswordScene.enter(async (ctx) => {
-  const { dialogMessageId } = ctx.session.postme;
-  const chatId = ctx.chat?.id;
+  try {
+    const { dialogMessageId } = ctx.session.postme;
+    const chatId = ctx.chat?.id;
 
-  await ctx.telegram.editMessageText(
-    chatId,
-    dialogMessageId,
-    undefined,
-    'Повторите пароль еще раз',
-    {
-      reply_markup: {
-        inline_keyboard: [
-          [{ text: 'Выйти', callback_data: CommonActions.ExitCallback }],
-        ],
-      },
-    },
-  );
-});
-
-confirmPasswordScene.on(message('text'), async (ctx) => {
-  const { password, dialogMessageId } = ctx.session.postme;
-  const { id: chatId } = ctx.chat;
-  const confirmPassword = ctx.message.text;
-
-  await ctx.deleteMessage();
-
-  if (password !== confirmPassword) {
     await ctx.telegram.editMessageText(
       chatId,
       dialogMessageId,
       undefined,
-      'Пароли не совпадают, попробуйте еще раз',
+      'Повторите пароль еще раз',
       {
         reply_markup: {
           inline_keyboard: [
@@ -128,18 +114,48 @@ confirmPasswordScene.on(message('text'), async (ctx) => {
         },
       },
     );
-    return;
+  } catch (error) {
+    console.log(error);
   }
+});
 
-  await ctx.deleteMessage(dialogMessageId);
+confirmPasswordScene.on(message('text'), async (ctx) => {
+  try {
+    const { password, dialogMessageId } = ctx.session.postme;
+    const { id: chatId } = ctx.chat;
+    const confirmPassword = ctx.message.text;
 
-  const errorMessage = await addChatAsResource(chatId, password);
+    await ctx.deleteMessage();
 
-  if (errorMessage) {
-    await ctx.reply(errorMessage);
-    return;
+    if (password !== confirmPassword) {
+      await ctx.telegram.editMessageText(
+        chatId,
+        dialogMessageId,
+        undefined,
+        'Пароли не совпадают, попробуйте еще раз',
+        {
+          reply_markup: {
+            inline_keyboard: [
+              [{ text: 'Выйти', callback_data: CommonActions.ExitCallback }],
+            ],
+          },
+        },
+      );
+      return;
+    }
+
+    await ctx.deleteMessage(dialogMessageId);
+
+    const errorMessage = await addChatAsResource(chatId, password);
+
+    if (errorMessage) {
+      await ctx.reply(errorMessage);
+      return;
+    }
+    ctx.reply('Чат успешно добавлен в базу!');
+  } catch (error) {
+    console.log(error);
   }
-  ctx.reply('Чат успешно добавлен в базу!');
 });
 
 // Main
@@ -149,62 +165,9 @@ export const checkPasswordScene = new Scenes.BaseScene<BotContext>(
 );
 
 checkPasswordScene.enter(async (ctx) => {
-  const editedMessage = await ctx.editMessageText(
-    'Чат защищен паролем, ведите пароль чтобы получить доступ:',
-    {
-      reply_markup: {
-        inline_keyboard: [
-          [{ text: 'Выйти', callback_data: CommonActions.ExitCallback }],
-        ],
-      },
-    },
-  );
-  if (typeof editedMessage === 'object') {
-    ctx.scene.state = {
-      ...ctx.scene.state,
-      dialogMessageId: editedMessage.message_id,
-    };
-  }
-});
-
-checkPasswordScene.on(message('text'), async (ctx) => {
-  const { id: chatId } = ctx.chat;
-  const {
-    dialogMessageId,
-    recourceChatId,
-    attemptions = 1,
-  } = ctx.scene.state as SceneState;
-
-  if (!dialogMessageId || !recourceChatId) {
-    await ctx.editMessageText(WRONG);
-    await ctx.scene.leave();
-    return;
-  }
-
-  if (attemptions === 3) {
-    await ctx.deleteMessage(dialogMessageId);
-    await ctx.reply('Превышен лимит, попробуйте позже');
-    ctx.scene.leave();
-    return;
-  }
-
-  const { text } = ctx.message;
-
-  await ctx.deleteMessage();
-
-  const isValidPassword = await comparePassword(text, recourceChatId);
-
-  if (!isValidPassword) {
-    ctx.scene.state = {
-      ...ctx.scene.state,
-      attemptions: attemptions + 1,
-    } as SceneState;
-
-    await ctx.telegram.editMessageText(
-      chatId,
-      dialogMessageId,
-      undefined,
-      `Неверный пароль, у вас осталось ${3 - attemptions} попыток`,
+  try {
+    const editedMessage = await ctx.editMessageText(
+      'Чат защищен паролем, ведите пароль чтобы получить доступ:',
       {
         reply_markup: {
           inline_keyboard: [
@@ -213,16 +176,81 @@ checkPasswordScene.on(message('text'), async (ctx) => {
         },
       },
     );
-    return;
+    if (typeof editedMessage === 'object') {
+      ctx.scene.state = {
+        ...ctx.scene.state,
+        dialogMessageId: editedMessage.message_id,
+      };
+    }
+  } catch (error) {
+    console.log(error);
   }
+});
 
-  await ctx.scene.leave();
+checkPasswordScene.on(message('text'), async (ctx) => {
+  try {
+    const { id: chatId } = ctx.chat;
+    const { id: userId } = ctx.from;
+    const {
+      dialogMessageId,
+      recourceChatId,
+      attemptions = 1,
+    } = ctx.scene.state as SceneState;
 
-  const isSuccess = await setResourceToListening(chatId, recourceChatId);
+    if (!dialogMessageId || !recourceChatId) {
+      await ctx.editMessageText(WRONG);
+      await ctx.scene.leave();
+      return;
+    }
 
-  await ctx.deleteMessage(dialogMessageId);
+    if (attemptions === 3) {
+      await ctx.deleteMessage(dialogMessageId);
+      await ctx.reply('Превышен лимит, попробуйте позже');
+      ctx.scene.leave();
+      return;
+    }
 
-  await ctx.reply(isSuccess ? 'Чат успешно выбран' : 'Невозможно выбрать чат');
+    const { text } = ctx.message;
+
+    await ctx.deleteMessage();
+
+    const isValidPassword = await comparePassword(text, recourceChatId);
+
+    if (!isValidPassword) {
+      ctx.scene.state = {
+        ...ctx.scene.state,
+        attemptions: attemptions + 1,
+      } as SceneState;
+
+      await ctx.telegram.editMessageText(
+        chatId,
+        dialogMessageId,
+        undefined,
+        `Неверный пароль, у вас осталось ${3 - attemptions} попыток`,
+        {
+          reply_markup: {
+            inline_keyboard: [
+              [{ text: 'Выйти', callback_data: CommonActions.ExitCallback }],
+            ],
+          },
+        },
+      );
+      return;
+    }
+
+    await ctx.scene.leave();
+
+    const isSuccess = await setResourceToListening(userId, recourceChatId);
+
+    await ctx.deleteMessage(dialogMessageId);
+
+    await ctx.reply(
+      isSuccess ? 'Чат успешно выбран' : 'Невозможно выбрать чат',
+    );
+  } catch (error) {
+    console.log(error);
+    await ctx.scene.leave();
+  }
 });
 
 export default [enterPasswordScene, confirmPasswordScene, checkPasswordScene];
